// types/property.ts
export type Property = {
  id: string;
  propertyName: string;
  neighborhood: string;
  street: string;
  city: string;
  priceNIS: number;
  beds: number;
  baths: number;
  indoorSqm: number;
  outdoorSqm: number;
  brokerName: string;
  brokerPhone: string;
  brokerEmail: string;
  heroImageUrl: string;
  backdropImageUrl: string;
  thumbImageUrl: string;
};